SSCS-7639 Code dup
